chore: drop the -cli suffix from paths-cli, gates-cli, and findings-cli
Every bundle command is already a CLI, so the suffix leaks (#615). The
`paths`, `gates`, and `findings` commands replace `paths-cli`,
`gates-cli`, and `findings-cli` in the age, cook, and cure bundles.

Following the `handoff` precedent from #621, each `*_cli.py` wrapper
merges into its library module (`shared/paths.py`, `shared/gates.py`,
`shared/findings.py`) so one module owns both the rules and the CLI.
Skill prose, generated command tables, and the CLI tests follow the
new names. A bundle test now rejects any command name carrying the
`-cli` suffix so it cannot creep back.

Closes #615
