+++ src/easy_cheese/shared/findings.py
@@ -15,8 +15,13 @@ in lockstep.
 
 from __future__ import annotations
 
+import argparse
 import re
 from dataclasses import dataclass, field
+from pathlib import Path
+from typing import TextIO, cast
+
+from easy_cheese.shared import cli
 
 SEVERITIES: tuple[str, ...] = ("blocker", "high", "medium", "low")
 SEVERITY_ORDER = {sev: i for i, sev in enumerate(SEVERITIES)}
@@ -292,3 +297,57 @@ def _split_composed_verb(verb: str) -> list[str]:
     if _NUM_LIST_RE.match(verb.replace(" ", "")):
         return [verb]
     return [piece.strip() for piece in verb.split(",") if piece.strip()]
+
+
+# ---- CLI: render-table, parse-selection ----
+def _load_findings(report_path: str) -> list[Finding]:
+    path = Path(report_path)
+    if not path.is_file():
+        raise cli.CliError(f"report not found: {report_path}")
+    return parse_findings_report(path.read_text(encoding="utf-8"))
+
+
+def _cmd_render_table(args: argparse.Namespace) -> None:
+    items = _load_findings(cast(str, args.report))
+    table = render_selection_table(items)
+    cli.emit(
+        table,
+        full=cast(bool, args.full),
+        json_mode=cast(bool, args.json_mode),
+        stdout=cast("TextIO", args.stdout),
+    )
+
+
+def _cmd_parse_selection(args: argparse.Namespace) -> None:
+    items = _load_findings(cast(str, args.report))
+    try:
+        ids = parse_selection(cast(str, args.selection), items)
+    except SelectionError as exc:
+        raise cli.CliError(str(exc)) from exc
+    cli.emit(
+        ids,
+        full=cast(bool, args.full),
+        json_mode=cast(bool, args.json_mode),
+        stdout=cast("TextIO", args.stdout),
+    )
+
+
+def _setup(parser: argparse.ArgumentParser) -> None:
+    sub = parser.add_subparsers(dest="cmd", required=True)
+
+    render = sub.add_parser("render-table", help="render selection table from an /age report")
+    _ = render.add_argument("--report", required=True, help="path to /age findings report")
+    render.set_defaults(func=_cmd_render_table)
+
+    select = sub.add_parser("parse-selection", help="resolve a selection verb to finding ids")
+    _ = select.add_argument("--report", required=True, help="path to /age findings report")
+    _ = select.add_argument("--selection", required=True, help="selection verb (e.g. 'all-high', '1,3', 'skip 2')")
+    select.set_defaults(func=_cmd_parse_selection)
+
+
+def main(argv: list[str]) -> int:
+    return cli.run(_setup, argv=argv)
+
+
+if __name__ == "__main__":
+    raise SystemExit(cli.run(_setup))
